feat(admin_tools): categoryheader adapted for topics + category
