+++ sefaria/urls.py
@@ -258,6 +258,7 @@ urlpatterns += [
     url(r'^api/topics/(?P<topic>.+)$', reader_views.topics_api),
     url(r'^api/topic/new$', reader_views.add_new_topic_api),
     url(r'^api/topic/delete/(?P<topic>.+)$', reader_views.delete_topic),
+    url(r'^api/topics/reorder$', reader_views.reorder_topics),
     url(r'^api/bulktopics$', reader_views.bulk_topic_api),
     url(r'^api/recommend/topics(/(?P<ref_list>.+))?', reader_views.recommend_topics_api),
 ]
