+++ static/js/TopicsPage.jsx
@@ -13,11 +13,11 @@ import Sefaria  from './sefaria/sefaria';
 import $  from './sefaria/sefariaJquery';
 import { NavSidebar, Modules } from './NavSidebar';
 import Footer  from './Footer';
+import {CategoryHeader} from "./Misc";
 import Component from 'react-class';
 
 // The root topics page listing topic categories to browse
 const TopicsPage = ({setNavTopic, multiPanel, initialWidth}) => {
-  const [addingTopics, toggleAddingTopics] = useEditToggle();
   let categoryListings = Sefaria.topic_toc.map(cat => {
     const openCat = e => {e.preventDefault(); setNavTopic(cat.slug, {en: cat.en, he: cat.he})};
     return (
@@ -58,15 +58,7 @@ const TopicsPage = ({setNavTopic, multiPanel, initialWidth}) => {
     {type: "GetTheApp"},
     {type: "SupportSefaria"},
   ];
-  let topicStatus = null;
 
-  if (Sefaria.is_moderator && addingTopics) {
-      const origData = {origEn: ""};
-      topicStatus = <TopicEditor close={toggleAddingTopics} origData={origData} onCreateSuccess={(slug) => window.location.href = "/topics/" + slug}/>;
-  }
-  else if (Sefaria.is_moderator) {
-      topicStatus = <AdminEditorButton text="Create a Topic" toggleAddingTopics={toggleAddingTopics}/>;
-  }
 
   return (
     <div className="readerNavMenu noLangToggleInHebrew" key="0">
@@ -74,8 +66,7 @@ const TopicsPage = ({setNavTopic, multiPanel, initialWidth}) => {
         <div className="sidebarLayout">
           <div className="contentInner">
               <div className="navTitle tight sans-serif">
-                <h1 className="sans-serif"><InterfaceText>Explore by Topic</InterfaceText></h1>
-                {topicStatus}
+                  <CategoryHeader title="Explore by Topic" path={[]} type="topics"/>
               </div>
               { about }
               { categoryListings }
